DtcDerivedPlaceholderModelとDtcTransformModelのインターフェースを追加し、実装クラスを作成。DtcLogicImplでの使用を更新。
