+++ src/main/java/kmg/tool/domain/model/io/dtc/DtcTransformModel.java
@@ -1,8 +1,5 @@
 package kmg.tool.domain.model.io.dtc;
 
-import kmg.core.infrastructure.type.KmgString;
-import kmg.tool.domain.types.io.dtc.DtcTransformTypes;
-
 /**
  * テンプレートの動的変換モデル<br>
  *
@@ -10,76 +7,24 @@ import kmg.tool.domain.types.io.dtc.DtcTransformTypes;
  *
  * @sine 1.0.0
  */
-public class DtcTransformModel {
-
-    /** 元の値 */
-    private final String originalValue;
-
-    /** 変換処理の種類 */
-    private DtcTransformTypes dtcTransformTypes;
-
-    /** 変換後の値 */
-    private String transformedValue;
-
-    /**
-     * コンストラクタ<br>
-     *
-     * @param value
-     *                          元の値
-     * @param dtcTransformTypes
-     *                          変換処理の種類
-     */
-    public DtcTransformModel(final String value, final DtcTransformTypes dtcTransformTypes) {
-
-        this.dtcTransformTypes = dtcTransformTypes;
-
-        if (value == null) {
-
-            this.originalValue = KmgString.EMPTY;
-            this.transformedValue = KmgString.EMPTY;
-            return;
-
-        }
-
-        this.originalValue = value;
-        this.transformedValue = value;
-
-    }
+public interface DtcTransformModel {
 
     /**
      * 指定された変換処理を適用する<br>
      */
-    public void apply() {
-
-        this.transformedValue = this.dtcTransformTypes.transform(this.originalValue);
-
-    }
+    void apply();
 
     /**
      * 元の値を返す<br>
      *
      * @return 元の値
      */
-    public String getOriginalValue() {
-
-        String result;
-
-        result = this.originalValue;
-        return result;
-
-    }
+    String getOriginalValue();
 
     /**
      * 変換後の値を返す<br>
      *
      * @return 変換後の値
      */
-    public String getTransformedValue() {
-
-        String result;
-
-        result = this.transformedValue;
-        return result;
-
-    }
+    String getTransformedValue();
 }
